spi: rename mod util to cs
