+++ boards/nordic/nrf52840dk/src/lib.rs
@@ -697,7 +697,7 @@ pub unsafe fn start() -> (
     let spi_controller = components::spi::SpiSyscallComponent::new(
         board_kernel,
         mux_spi,
-        kernel::hil::spi::util::IntoChipSelect::<_, kernel::hil::spi::util::ActiveLow>::into_cs(
+        kernel::hil::spi::cs::IntoChipSelect::<_, kernel::hil::spi::cs::ActiveLow>::into_cs(
             &gpio_port[SPI_CS],
         ),
         capsules_core::spi_controller::DRIVER_NUM,
